+++ HISTORY.md
@@ -11,6 +11,7 @@
 ### Bugfixes
 
 - Fixed minimal openpyxl dependency version to 2.6.0 (#457).
+- Dates from xls files are now read as Python datetime objects (#373).
 
 ## 1.1.0 (2020-02-13)